requirements.txt
anthropic
requests
beautifulsoup4
PyYAML
